requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bg-helper==0.1.23
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
dt-helper==0.0.10
exceptiongroup==1.2.2
fastapi==0.115.12
fs-helper==0.0.11
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
input-helper==0.1.53
mongo-helper==0.0.13
motor==3.7.0
neo4j==5.28.1
numpy==2.2.4
pandas==2.2.3
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
settings-helper==0.0.14
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uvicorn==0.34.0